Implementation of dataToString method This is (rudimentarily) used to display the values of all statistics stored.
It does display some non-relevant attributes and is not a clean solution, however it suffices to display that all attributes are correctly saved, and can be called upon.

It would be trivial but very time consuming to adapt this function to return the actual stats such as % points won etc.